% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_nbd.R
\name{bg_nbd}
\alias{bg_nbd}
\title{Beta Geometric/Negative Binomial Distrubution Model}
\usage{
bg_nbd(rfm_data, estimate = FALSE)
}
\arguments{
\item{rfm_data}{\preformatted{A data frame of rfm formatted data. Columns should refer to:
x: repeat purchases
t_x: time of last repeat
T_: calibration period}}

\item{estimate}{Is this for individual log likelihood contributions or for estimating the
parameters of the distribution?}
}
\value{
\preformatted{
 Inner function does the work of estimating the parameters of the bg/nbd
 model. Returns either the individual log likelihood contributions or
 negative log likelihood for estimating the bg/nbd parameters.
 r: shape parameter of gamma distribution
 alpha: scale parameter of gamma distribution
 a: beta function shape parameter
 b: beta function shape parameter}
}
\description{
A functional implementation of the bg/nbd model for rfm analysis.
Implemented from the paper: “Counting Your Customers” the Easy Way: An Alternative to the Pareto/NBD Model
\url{http://www.brucehardie.com/papers/bgnbd_2004-04-20.pdf}
Functions are adapted from:
\url{http://www.brucehardie.com/notes/004/bgnbd_spreadsheet_note.pdf}
}
\details{
Outer function that sets the environment.
}
