% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDNOW_sample_summary}
\alias{CDNOW_sample_summary}
\title{CDNOW Summary Data}
\format{
A data frame with 2357 rows and 4 variables:
\describe{
\item{sample_id}{}
\item{x}{repeat purchases}
\item{t_x}{time of last repeat}
\item{T_}{calibration period}
}
}
\source{
\url{http://www.brucehardie.com/notes/026/notes_on_CDNOW_master.pdf}
}
\usage{
CDNOW_sample_summary
}
\description{
CDNOW training summary data set
}
\keyword{datasets}
