% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{.gh}
\alias{.gh}
\title{Gaussian hypergeometric function}
\usage{
.gh(r, alpha, a, b, x, T_, t)
}
\arguments{
\item{r}{shape parameter of gamma distribution}

\item{alpha}{scale parameter of gamma distribution}

\item{a}{beta function shape parameter}

\item{b}{beta function shape parameter}

\item{x}{repeat purchases}

\item{T_}{calibration period}

\item{t}{forecast window}
}
\description{
Gaussian hypergeometric function
}
\keyword{internal}
