% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Conditional Expectation}
\usage{
forecast(r = 0.2426707, alpha = 4.4159851, a = 0.7928716, b = 2.4247084)
}
\arguments{
\item{r}{shape parameter of gamma distribution}

\item{alpha}{scale parameter of gamma distribution}

\item{a}{beta function shape parameter}

\item{b}{beta function shape parameter}
}
\value{
\preformatted{
 Inner function does the work of estimating the parameters of the bg/nbd
 model. Returns the forecasted number of transactions.
 x: repeat purchases
 t_x: time of last repeat
 T_: calibration period
 t: forecast window}
}
\description{
Returns the expected number of transactions for a given forecast
history
}
